+++ ui/components/auth/oss/sign-in-form.tsx
@@ -156,7 +156,6 @@ export const SignInForm = ({
             type="email"
             label="Email"
             placeholder="Enter your email"
-            showFormMessage
           />
           {!isSamlMode && (
             <CustomInput control={form.control} name="password" password />